Project_docs/document upload deck/requirements.txt
# requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
# psycopg2==2.9.10
cryptography==42.0.2
PyJWT==2.8.0
python-dotenv==1.0.1
Werkzeug==3.1.3
polars==0.19.19
reportlab==4.0.8
SQLAlchemy==2.0.36
pytest==8.3.4
pytest-mock==3.14.0
schedule==1.2.2
setuptools==75.8.0
# Removed functools as it's built into Python
# Removed pathlib as it's built into Python
redis==5.2.1
psycopg2-binary==2.9.10
tabulate==0.9.0
requests==2.32.3
flask-wtf==1.2.2
selenium==4.29.0
webdriver_manager==4.0.2
email_validator==2.2.0          
debugpy==1.8.13
twilio==9.5.1
Pillow==11.1.0